requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.0
fire==0.5.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-core==1.2.7
langchain-experimental==0.0.37
langchain-openai==1.1.7
langcorn==0.0.22
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langserve==0.3.3
langsmith==0.6.6
loguru==0.7.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numexpr==2.14.1
numpy==1.26.4
openai==2.16.0
orjson==3.11.6
ormsgpack==1.12.2
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sse-starlette==1.8.2
starlette==0.50.0
tenacity==8.5.0
termcolor==3.3.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
